Feature: Service Sync (#47)
This PR is a preparation PR for the followup PR that will implement
Service Sync feature. The purpose of this feature is to facilitate
exposing a service in one cluster to other clusters, using Cilium Global
Services, available on PAC-v2 clusters.

This PR adds a doc about that feature, and also performs a bunch of
small fixes and housekeeping tasks that will cleanup the next PR. These
various changes:

- Fixes a few bugs that affect running tests already present on this
repo
- Adds Cilium to the local test clusters. Cilium is a required component
for the new feature.
- Adds some example files useful for testing
- Merges pending renovate PRs

All of this does not change regular operation of keess and this changes
need not to be deployed to production.

See
[docs/service-sync.md](https://github.com/powerhome/keess/blob/sync-services/docs/service-sync.md)
in this PR for more context and for the design of the new feature.

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>