(feat) Allow chart to disable delete permissions from Keess (#75)
## 🚀 What does this PR do?

Adding chart option to disable the delete permission given to Keess
Service Account.

As a bonus, it adds:
- A PR template to this repo (this PR is using it)
- Docs about Keess versioning (I made a decision on how to change the
chart version, and described it here)

## 🔍 Why? (Context)
<!--
- (REQUIRED) Why may be MORE important than what.
- Help your reviewer friend and explain the why and add some context
about the change.
-->

This is a failsafe measure to be used when making changes or operations
that we fear could have a chance of mistakenly deleting objects as
orphans. We can remove the permissions Keess uses to delete those
objects, preventing it from doing those mistakes.

## Checklist

<!--
this is here to help you REMEMBER to:
-->
- [x] Update the Helm chart version and/or appVersion if those are
changed
- [ ] Update cmd/version.go if code was changed
- [ ] Publish the new keess image
- [ ] Publish the new kubeconfig-reloader image
- [x] Updated docs/ or /README.md if needed

## 🧪 How this was tested? And test results
<!--
- Remember to test on alpha, unless it does not apply (docs, workflows,
etc.).
- How can the reviewer test it? Commands, etc.
- Also add test evidences if possible, showing it works.
-->

Tested on alpha.

## ⚠️ Special care or coordination needed?
<!--
- Do we need to coordinate merge with other PRs? List them here.
- Do we need to coordinate or add changes to external resources (F5,
AWS, etc?) mention them here
- Is it a risk operation that may cause service interruption? Do we need
an implementation plan?
-->

`No special care needed.`