Fix go routines return bugs (#86)
## 🚀 What does this PR do?
<!-- (REQUIRED) Start with what is being done here -->

This fixes #78. Poller functions had a "return" inside the go routine
function, that caused the go routine to exit when facing some transient
error polling the kubernetes api.

This also adds unit test cases to catch that problems if we reintroduce
it by mistake.

## 🔍 Why? (Context)
<!--
- (REQUIRED) Why may be MORE important than what.
- Help your reviewer friend and explain the why and add some context
about the change.
-->

With new metrics, alerts and log messages introduced by keess 1.4.0, we
were able to confirm the go routines are exiting, and correlate clearly
with the errors listing objects on kube api.

## Checklist

<!--
this is here to help you REMEMBER to:
-->
- [x] Update the Helm chart version and/or appVersion if those are
changed
- [x] Update cmd/version.go if code was changed
- [x] Publish the new keess image
- [x] Publish the new kubeconfig-reloader image
- [x] Updated docs/ or /README.md if needed

## 🧪 How this was tested? And test results
<!--
- Remember to test on alpha, unless it does not apply (docs, workflows,
etc.).
- How can the reviewer test it? Commands, etc.
- Also add test evidences if possible, showing it works.
-->

Tested on alpha.

But as for the bug being fixed, I:

- Run the new unit tests with and without the fixing commit, making sure
they fail without the fixes.

## ⚠️ Special care or coordination needed?
<!--
- Do we need to coordinate merge with other PRs? List them here.
- Do we need to coordinate or add changes to external resources (F5,
AWS, etc?) mention them here
- Is it a risk operation that may cause service interruption? Do we need
an implementation plan?
-->

`No special care needed.`